chore: setup project app & server
Frontend will be built using React Native (expo cli). It uses JavaScript for now
and then will use TypeScript in the near future.

Backend will be built using nodejs (express server) with TypeSript.

This patch included the setup for both the app and the server. We are using yarn
workspaces to have a monorepo for both the app an the server.